spark-expectations-demo/requirements.txt
pyspark==3.5.0
spark-expectations==2.3.1
delta-spark==3.0.0
Jinja2==3.1.6